Fixes problem when calling read after an error
I'm not precisely sure what happens, but I have some user code that I think is calling read after there is an error reading (perhaps the previous read was partially satisfied by the peek buffer). In this case, I believe `DATA(io)->length` is 0 (or -1) which causes some scary maths to happen, and then a memcpy of a negative number of bytes.