Update zstd configure check to use AC_CHECK_LIB
Ubuntu 16.04 has a version of libzstd that does not have some
of the functions that we use, so we need to NOT build zstd
support in that case -- the old check would pass whenever
zstd.h was present even if it was missing key API functions.

Also fixed a couple of copy-paste errors in the zstd checks as
well.