Fix bug where lz4 decoding would fail with srcPtr_wrong error.
Older versions of liblz4 throw an error if the source buffer
address changes during decompression -- however, we sometimes
use memmove to free up space in our internal read buffer before
reading the next lot of compressed data from the input file.
This will change the address of the "src ptr" from wherever
it used to be in the buffer to the start of the buffer and cause
the error to be thrown.

Later versions of liblz4 are not so ridiculously strict about
where the next bit of input comes from.

The hack fix here is to check for LZ4F_getVersion() function at
configure time -- this function was added in the version that
removed the error. We now only do a memmove if the check passed.
If not, we try to carry on without the memmove and hope we don't
get stuck in a situation where we have no space in our input
buffer yet not enough unprocessed content to consume what is left
in there.