Add separate configure check for lz4 streaming API
We plan to use the streaming API as a fallback option if the
frame API is not present (which occurs in liblz4-dev on both
xenial and jessie).

Updated existing lz4 code to correctly check the new
HAVE_LIBLZ4F #define which is set if the frame API is available.