iow-thread: avoid zero-length writes when closing
In thread_consumer, when we exit the inner loop because the program is over,
the current buffer may be completely empty.  In that case, we perform a
zero-length call to wandio_wwrite, where an uninitialized read occurs.

Since the last buffer might be partially filled, we can't just jump out of
the outer loop.  Instead, just make sure to avoid calling wandio_wwrite with
a completely empty buffer.

Found with Valgrind.