Add DLLEXPORT to a *_open and *_wopen functions
These functions were already in the wandio.h header and some
users would like to access them (the HTTP ones, in particular),
directly without the extra wandio overhead. Without DLLEXPORT,
the symbols are only available internally so this commit fixes
that.

Long term, we need to re-evaluate the structure of wandio and
how things should be arranged / made available to best suit the
modern use cases that people have for it.